backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
python-dotenv>=1.0.0
python-multipart>=0.0.6
pydantic>=1.10.7
httpx>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
redis>=4.5.5
langchain>=0.0.335
langchain-openai>=0.0.1
openai>=1.0.0
twilio>=8.9.0
litellm>=1.0.0
supabase>=2.0.0
pytz>=2023.3
python-dateutil>=2.8.2